move tile peptides tests to new file
